Made getCacheData re-usable
